+++ app/src/main/cpp/jni_bridge.h
@@ -2,4 +2,4 @@
 
 #include <vector>
 
-int requestSamples(int numSamples, float *outputData);
+int requestSamples(int numSamples, int16_t *outputData);
